# roomix-backend/src/main/java/com/roomix/api/model/dto/request/LoginRequest.java
package com.roomix.api.model.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class LoginRequest {

    @NotBlank
    @Email
    private String email;

    @NotBlank
    private String password;
}